fix mistakes in porting SMART science in the components
I compared standalone SMART outputs with outputs using the SMART components in `cm4twc`, but the hydrographs were quite different. With this commit, hydrographs are very similar (not perfect match because of the concurrent coupling in `cm4twc`, while it is sequential in SMART standalone, and also because of the slight change in science in surfacelayer), however, the river flow is off by a factor of 10^-3 in the one obtained using `cm4twc`.

So there is still this factor to be fixed in a later commit. It could have been a difference of 1000 in the inputs, since I needed to get a rainfall_flux in kg m-2 s-1 from data in mm day-1), but I double-checked, and I seem to have done it correctly.